+++ docs/changelog.rst
@@ -8,8 +8,12 @@ CHANGELOG
 **Bug fixes**
 
 - Fix migration 2.77.1 publication
+
+**Minor improvements**
+
 - Add publication informations by lang on infrastructure
 - Remove table Infrastructure on infrastructure
+- Add translation signage
 
 
 2.77.1     (2022-03-11)