+++ docs/changelog.rst
@@ -10,6 +10,7 @@ CHANGELOG
 - Fix migration 2.77.1 publication
 - Add publication informations by lang on infrastructure
 - Remove table Infrastructure on infrastructure
+- Fix Intervention detail page breaks when target is a Report
 
 
 2.77.1     (2022-03-11)