+++ docs/changelog.rst
@@ -8,9 +8,13 @@ CHANGELOG
 **Bug fixes**
 
 - Fix migration 2.77.1 publication
+
+**Minor improvements**
+
 - Add publication informations by lang on infrastructure
 - Remove table Infrastructure on infrastructure
 - Fix Intervention detail page breaks when target is a Report
+- Add translation signage
 
 
 2.77.1     (2022-03-11)